refactor and make some fixes
