+++ Dockerfile
@@ -8,7 +8,7 @@ RUN pip install --no-cache-dir -r requirements.txt
 RUN useradd --create-home appuser
 USER appuser
 
-COPY server.py services.py registrations.py giving.py calendar_events.py .
+COPY app.py server.py services.py registrations.py giving.py calendar_events.py .
 
 EXPOSE 8000
