+++ giving.py
@@ -3,7 +3,7 @@ import asyncio
 from fastmcp.dependencies import Depends
 from pypco import PCO
 
-from server import mcp, get_pco, _build_patch_body
+from app import mcp, get_pco, _build_patch_body
 
 
 # =============================================================================
